[cleanup] Detect target environment
1. Detect the target environment (in `macros.h`) by examining compiler
   defined preprocessor symbols.  Define one or more of
   `OT_PLATFORM_TARGET`, `OT_PLATFORM_RV32` and `OT_PLATFORM_HOST` depending
   on the compiler target.
2. Eliminate use of MOCK_*_MMIO in favor of OT_PLATFORM_*.  The mmio and other
   target specific libraries will never work on the host (unittest) platform.
   The only versions of such libraries which make sense are the mocked
   libraries.
3. Relax the `target_compatible_with` constraints on many of the Mask ROM's
   `cc_library` targets.  Being compatible only with OPENTITAN_CPU causes
   many of the unittests to be skipped.
