+++ sw/device/lib/base/testing/mock_abs_mmio.h
@@ -8,7 +8,7 @@
 #include "sw/device/lib/base/abs_mmio.h"
 #include "sw/device/lib/base/testing/global_mock.h"
 #include "sw/device/lib/base/testing/mock_mmio_test_utils.h"
-#include "sw/device/silicon_creator/testing/mask_rom_test.h"
+//#include "sw/device/silicon_creator/testing/mask_rom_test.h"
 
 namespace mask_rom_test {
 namespace internal {
@@ -98,33 +98,6 @@ using MockAbsMmio = testing::StrictMock<internal::MockAbsMmio>;
   EXPECT_CALL(::mask_rom_test::MockAbsMmio::Instance(), \
               Write32Shadowed(addr, mock_mmio::ToInt<uint32_t>(__VA_ARGS__)));
 
-extern "C" {
-
-uint8_t abs_mmio_read8(uint32_t addr) {
-  return MockAbsMmio::Instance().Read8(addr);
-}
-
-void abs_mmio_write8(uint32_t addr, uint8_t value) {
-  MockAbsMmio::Instance().Write8(addr, value);
-}
-
-void abs_mmio_write8_shadowed(uint32_t addr, uint8_t value) {
-  MockAbsMmio::Instance().Write8Shadowed(addr, value);
-}
-
-uint32_t abs_mmio_read32(uint32_t addr) {
-  return MockAbsMmio::Instance().Read32(addr);
-}
-
-void abs_mmio_write32(uint32_t addr, uint32_t value) {
-  MockAbsMmio::Instance().Write32(addr, value);
-}
-
-void abs_mmio_write32_shadowed(uint32_t addr, uint32_t value) {
-  MockAbsMmio::Instance().Write32Shadowed(addr, value);
-}
-
-}  // extern "C"
 }  // namespace mask_rom_test
 
 #endif  // OPENTITAN_SW_DEVICE_LIB_BASE_TESTING_MOCK_ABS_MMIO_H_
