+++ sw/device/lib/base/testing/mock_mmio_test_utils.h
@@ -14,7 +14,6 @@
 
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
-#include "sw/device/lib/base/mmio.h"
 
 namespace mock_mmio {
