+++ sw/device/silicon_creator/lib/base/sec_mmio_unittest.cc
@@ -11,6 +11,7 @@
 #include "gtest/gtest.h"
 #include "sw/device/lib/base/testing/mock_abs_mmio.h"
 #include "sw/device/silicon_creator/lib/error.h"
+#include "sw/device/silicon_creator/testing/mask_rom_test.h"
 
 namespace sec_mmio_unittest {
 namespace {
